+++ CHANGES.md
@@ -82,6 +82,7 @@ Changes in 12.5
 - Use MultipleCheckboxInput() function in AddAbsences.php
 - CSP add .onclick-checkall CSS class & use `data-name-like` attribute in AddAbsences.php, DuplicateAttendance.php, Exceptions.php & Profiles.php
 - CSP add .onclick-toggle CSS class & use `data-id` attribute in CreateParents.php, ReportCards.fnc.php & Transcripts.fnc.php
+- FinalGradeSave() can save COMMENT column in FinalGrades.inc.php
 
 Changes in 12.4.3
 -----------------